fix tiny typo
This is all very intriguing and I've been having a go with docker since your UseR talk and Carl's blog posts. Have you done much testing with windows and macs (likely what many people will bring to SC workshops)? I've been testing a bit with windows with my students and it hasn't been a smooth experience, I think we're going to stick with Lubuntu in regular virtualbox VM for the moment. If people are already working on Linux then docker seems like a good fit, scientists who are not programmers and work on other OSs the benefits are less obvious, compared to a VM that is provisioned by a shell script, [for example](https://gist.github.com/benmarwick/11204658). Or am I missing the point?